+++ docs/docs/api/package/Values.t.d.ts
@@ -1,0 +1,5 @@
+import { TemplateFunction } from '@dxos/plate';
+import { Input } from '../config.t.js';
+declare const template: TemplateFunction<Input>;
+export default template;
+//# sourceMappingURL=Values.t.d.ts.map