+++ packages/common/typings/src/misc.d.ts
@@ -5,3 +5,4 @@
 declare module 'buffer-json-encoding';
 declare module 'end-of-stream-promise';
 declare module 'signal-promise';
+declare module 'y-monaco';