+++ packages/experimental/kai/src/hooks/useAppState.tsx
@@ -2,19 +2,20 @@
 // Copyright 2022 DXOS.org
 //
 
-import React, { Context, Dispatch, SetStateAction, createContext, FC, ReactNode, useContext, useState } from 'react';
+import React, { Context, Dispatch, SetStateAction, FC, ReactNode, createContext, useContext, useState } from 'react';
+
+// TODO(burdon): Merge options.
 
 // TODO(burdon): Use reducer?
 // https://beta.reactjs.org/learn/scaling-up-with-reducer-and-context
 
 export type AppState = {
+  // Debug info.
+  debug?: boolean;
+  // Auto-create data.
+  demo?: boolean;
+  // UX state.
   showSidebar?: boolean;
-  dummy?: number;
-};
-
-export const defaultAppState: AppState = {
-  showSidebar: true,
-  dummy: 100
 };
 
 export type AppStateContextType = [AppState, Dispatch<SetStateAction<AppState>>];
@@ -23,11 +24,8 @@ export const AppStateContext: Context<AppStateContextType | undefined> = createC
   undefined
 );
 
-export const AppStateProvider: FC<{ children: ReactNode; initialState?: AppState }> = ({
-  children,
-  initialState = defaultAppState
-}) => {
-  const [state, setState] = useState<AppState>(initialState);
+export const AppStateProvider: FC<{ children: ReactNode; value?: AppState }> = ({ children, value }) => {
+  const [state, setState] = useState<AppState>(value ?? { showSidebar: true });
   return <AppStateContext.Provider value={[state, setState]}>{children}</AppStateContext.Provider>;
 };